+++ src/components/toolbar/Toolbar.tsx
@@ -247,6 +247,8 @@ export default function Toolbar() {
             <option value="DFA">DFA</option>
             <option value="NFA">NFA</option>
             <option value="ENFA">ε-NFA</option>
+            <option value="DPDA">DPDA</option>
+            <option value="NPDA">NPDA</option>
           </select>
         </div>
       </div>